修改统计值

+++ include/DancingMatrix.h
@@ -82,13 +82,13 @@ class DancingMatrix
         struct CCExperimentStats {
             uint64_t decCalls = 0, incCalls = 0;
             uint64_t merges = 0, splits = 0, treeEdge_cuts = 0, cc_decompose = 0;
-            long double V1 = 0, E1 = 0; // 减量式图规模统计
-            long double V2 = 0, E2 = 0; // 增量式图规模统计 
-            long double Vd = 0, Ed = 0; // 减量式更新顶点数、边数
-            long double Vi = 0, Ei = 0; // 增量式更新顶点数、边数 
+            uint64_t V1 = 0, E1 = 0; // 减量式图规模统计
+            uint64_t V2 = 0, E2 = 0; // 增量式图规模统计 
+            uint64_t Vd = 0, Ed = 0; // 减量式更新顶点数、边数
+            uint64_t Vi = 0, Ei = 0; // 增量式更新顶点数、边数 
             uint64_t enSamples = 0, enSum = 0;
             uint64_t enPositiveUpdates = 0;
-            long double enUpdateAverageSum = 0;
+            uint64_t enUpdateAverageSum = 0;
             uint64_t replacementSearchCalls = 0, replacementScanSteps = 0;
             uint64_t replacementEarlyBreaks = 0, replacementFullScans = 0;
